agent_template_86174/requirements.txt
```
python-dotenv
langchain-anthropic
langgraph
langchain-community
langchain-core
```